# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface WorkspaceData { [key: string]: unknown }

export interface ProtradersConfig {
  configured: boolean;
  loginConfigured: boolean;
  signupConfigured: boolean;
  publicAppId: string;
  partnerParam: string;
  campaign: string;
}

export type ProtradersPreflightExecutionMode = typeof ProtradersPreflightExecutionMode[keyof typeof ProtradersPreflightExecutionMode];


export const ProtradersPreflightExecutionMode = {
  DISABLED: 'DISABLED',
  DEMO: 'DEMO',
  BOTH: 'BOTH',
  LIVE_LOCKED: 'LIVE LOCKED',
} as const;

export interface ProtradersPreflight {
  productionBaseUrl: string;
  redirectUri: string;
  https: boolean;
  oauthClientConfigured: boolean;
  partnerTrackingConfigured: boolean;
  sessionSecretConfigured: boolean;
  publicAppConfigured: boolean;
  frontendConfigured: boolean;
  tradingEnabled: boolean;
  liveTradingEnabled: boolean;
  demoOnly: boolean;
  maxStake: number;
  maxDuration: number;
  allowedSymbols: string[];
  executionMode: ProtradersPreflightExecutionMode;
  readyForControlledLiveTest: boolean;
  readyForRealTrading: boolean;
}

export interface SessionStatus {
  authenticated: boolean;
  /** @nullable */
  expiresAt?: number | null;
}

export interface TrackEventInput {
  type?: string;
  path?: string;
}

export interface Analytics {
  visitors: number;
  registrations: number;
  oauthSuccesses: number;
  /** @nullable */
  fundedAccounts: number | null;
  note: string;
  ephemeral: boolean;
}

export type AccountAccountType = typeof AccountAccountType[keyof typeof AccountAccountType];


export const AccountAccountType = {
  demo: 'demo',
  real: 'real',
} as const;

export interface Account {
  authenticated: boolean;
  /** @nullable */
  balance: number | null;
  /** @nullable */
  currency: string | null;
  /** @nullable */
  loginid: string | null;
  accountType: AccountAccountType;
  /** @nullable */
  openPnl: number | null;
}

export type TradeInputContractType = typeof TradeInputContractType[keyof typeof TradeInputContractType];


export const TradeInputContractType = {
  CALL: 'CALL',
  PUT: 'PUT',
} as const;

/**
 * Required only for real-money accounts; ignored for demo accounts.
 */
export type TradeInputLiveConfirmation = typeof TradeInputLiveConfirmation[keyof typeof TradeInputLiveConfirmation];


export const TradeInputLiveConfirmation = {
  CONFIRM_LIVE_TRADE: 'CONFIRM_LIVE_TRADE',
} as const;

export interface TradeInput {
  symbol: string;
  contract_type: TradeInputContractType;
  /** @exclusiveMinimum 0 */
  stake: number;
  /** @minimum 1 */
  duration: number;
  /** Required only for real-money accounts; ignored for demo accounts. */
  live_confirmation?: TradeInputLiveConfirmation;
}

export interface TradeResult {
  ok: boolean;
  message: string;
  /** @nullable */
  contractId: string | null;
}

export type GetMarketCandlesParams = {
granularity?: number;
count?: number;
};