# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deriv/login:
    get:
      operationId: startDerivLogin
      tags: [protraders]
      summary: Start Deriv login
      responses:
        "302":
          description: Redirects to Deriv OAuth
        "503":
          description: OAuth is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deriv/signup:
    get:
      operationId: startDerivSignup
      tags: [protraders]
      summary: Start Deriv signup
      responses:
        "302":
          description: Redirects to Deriv OAuth registration
        "503":
          description: Signup is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /track:
    post:
      operationId: trackEvent
      tags: [protraders]
      summary: Record an ephemeral product event
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackEventInput"
      responses:
        "204":
          description: Event recorded
...
  /account:
    get:
      operationId: getAccount
      tags: [protraders]
      summary: Read the authenticated Deriv account
      responses:
        "200":
          description: Account data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades:
    post:
      operationId: createTrade
      tags: [protraders]
      summary: Open a controlled Deriv trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "200":
          description: Trade opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: Invalid trade parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Demo account required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Trading disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    TrackEventInput:
      type: object
      properties:
        type:
          type: string
        path:
          type: string
...
    Account:
      type: object
      properties:
        authenticated:
          type: boolean
        balance:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        loginid:
          type: ["string", "null"]
        accountType:
          type: string
          enum: [demo, real]
        openPnl:
          type: ["number", "null"]
      required:
        - authenticated
        - balance
        - currency
        - loginid
        - accountType
        - openPnl
    TradeInput:
      type: object
      properties:
        symbol:
          type: string
        contract_type:
          type: string
          enum: [CALL, PUT]
        stake:
          type: number
          exclusiveMinimum: 0
        duration:
          type: number
          minimum: 1
        live_confirmation:
          type: string
          enum: [CONFIRM_LIVE_TRADE]
          description: Required only for real-money accounts; ignored for demo accounts.
      required:
        - symbol
        - contract_type
        - stake
        - duration
    TradeResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        contractId:
          type: ["string", "null"]
      required:
        - ok
        - message
        - contractId