# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInputContractType } from './tradeInputContractType';
import type { TradeInputLiveConfirmation } from './tradeInputLiveConfirmation';

export interface TradeInput {
  symbol: string;
  contract_type: TradeInputContractType;
  /** @exclusiveMinimum 0 */
  stake: number;
  /** @minimum 1 */
  duration: number;
  /** Required only for real-money accounts; ignored for demo accounts. */
  live_confirmation?: TradeInputLiveConfirmation;
}